Improve code coverage and documentation #39
