+++ docker-compose.yml
@@ -9,7 +9,7 @@ web:
         - mysqldb
         - redis
     volumes:
-        - ./:/data:cached
+        - ./:/data
     working_dir: "/data"
     environment:
         DATABASE_HOST: mysqldb
@@ -26,7 +26,7 @@ cli:
         - mysqldb
         - redis
     volumes:
-        - ./:/data:cached
+        - ./:/data
     working_dir: "/data"
     environment:
         DATABASE_HOST: mysqldb
