--- docs/documentation.md
@@ -1,11 +1,0 @@
-# Documentation #
-
-[Installation](installation.md)
-
-[Configuration](configuration.md)
-
-[API endpoints](api.md)
-
-[Demo integration](demo.md)
-
-[Write your own integrations](integrations.md)
