+++ src/Yay/Bundle/IntegrationBundle/Tests/Configuration/Fixture/transform_from_file.yml
