+++ src/Yay/Bundle/IntegrationBundle/Tests/Configuration/Fixture/validate_empty_configuration.yml
