+++ src/Yay/Bundle/IntegrationBundle/Tests/Configuration/Fixture/validate_full_configuration.yml
