+++ src/Yay/Bundle/IntegrationBundle/Tests/Service/Fixture/install_entities_supports_action.yml
