+++ src/Yay/Bundle/IntegrationBundle/Tests/Service/Fixture/load_entities.yml
