+++ src/Yay/Component/HttpFoundation/Tests/Request/ParamConverter/DeserializeFieldConverterTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yay\Component\HttpFoundation\Test\Request\ParamConverter;
+namespace Yay\Component\HttpFoundation\Tests\Request\ParamConverter;
 
 use JMS\Serializer\SerializerBuilder;
 use JMS\Serializer\SerializerInterface;
