+++ src/Yay/Component/HttpFoundation/Tests/Request/ParamConverter/JsonFieldConverterTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Yay\Component\HttpFoundation\Test\Request\ParamConverter;
+namespace Yay\Component\HttpFoundation\Tests\Request\ParamConverter;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\HttpFoundation\Request;
 use PHPUnit\Framework\TestCase;
-use Yay\Component\HttpFoundation\Request\ParamConverter\HeaderFieldConverter;
+use Yay\Component\HttpFoundation\Request\ParamConverter\JsonFieldConverter;
 
-class HeaderFieldConverterTest extends TestCase
+class JsonFieldConverterTest extends TestCase
 {
     /**
      * @param string $converterName
@@ -17,7 +17,7 @@ class HeaderFieldConverterTest extends TestCase
      * @return ParamConverter
      */
     public function createConfiguration(
-        string $converterName = 'HeaderField',
+        string $converterName = 'JsonField',
         string $parameterName = 'foo',
         array $options = []
     ): ParamConverter {
@@ -45,7 +45,7 @@ class HeaderFieldConverterTest extends TestCase
     {
         $configuration = $this->createConfiguration();
 
-        $this->assertTrue((new HeaderFieldConverter())->supports($configuration));
+        $this->assertTrue((new JsonFieldConverter())->supports($configuration));
     }
 
     /**
@@ -53,46 +53,59 @@ class HeaderFieldConverterTest extends TestCase
      */
     public function param_converter_is_not_supported(): void
     {
-        $configuration = $this->createConfiguration('HeaderField2');
+        $configuration = $this->createConfiguration('JsonField2');
 
-        $this->assertFalse((new HeaderFieldConverter())->supports($configuration));
+        $this->assertFalse((new JsonFieldConverter())->supports($configuration));
     }
 
     /**
      * @test
      */
-    public function header_field_is_applied_as_attribute(): void
+    public function json_field_is_applied_as_attribute(): void
     {
         $configuration = $this->createConfiguration();
-        $request = Request::create('/', 'GET');
-        $request->headers->set('foo', 'bar');
+        $content = '{"foo": "bar"}';
+        $request = Request::create('/', 'POST', [], [], [], [], $content);
 
-        (new HeaderFieldConverter())->apply($request, $configuration);
+        (new JsonFieldConverter())->apply($request, $configuration);
         $this->assertEquals('bar', $request->attributes->get('foo'));
     }
 
     /**
      * @test
      */
-    public function header_field_is_applied_as_attribute_with_field(): void
+    public function json_field_is_applied_as_attribute_with_field(): void
     {
-        $configuration = $this->createConfiguration('HeaderField', 'foo', ['field' => 'baz']);
-        $request = Request::create('/', 'GET');
-        $request->headers->set('baz', 'bar');
+        $configuration = $this->createConfiguration('JsonField', 'foo', ['field' => 'baz']);
+        $content = '{"baz": "bar"}';
+        $request = Request::create('/', 'POST', [], [], [], [], $content);
 
-        (new HeaderFieldConverter())->apply($request, $configuration);
+        (new JsonFieldConverter())->apply($request, $configuration);
         $this->assertEquals('bar', $request->attributes->get('foo'));
     }
 
     /**
      * @test
      */
-    public function header_field_is_not_set(): void
+    public function json_field_is_not_set(): void
     {
         $configuration = $this->createConfiguration();
-        $request = Request::create('/', 'GET');
+        $request = Request::create('/', 'POST');
 
-        (new HeaderFieldConverter())->apply($request, $configuration);
+        (new JsonFieldConverter())->apply($request, $configuration);
+        $this->assertFalse($request->attributes->has('foo'));
+    }
+
+    /**
+     * @test
+     */
+    public function json_field_is_invalid(): void
+    {
+        $configuration = $this->createConfiguration();
+        $content = '{"baz": ';
+        $request = Request::create('/', 'POST', [], [], [], [], $content);
+
+        (new JsonFieldConverter())->apply($request, $configuration);
         $this->assertFalse($request->attributes->has('foo'));
     }
 }
